v1.1.0 chunk upload

+++ behaviors/FileUploadBehavior.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace bttree\smyupload\behaviors;
 
 use bttree\smyupload\components\UploadManager;
@@ -74,16 +75,39 @@ class FileUploadBehavior extends Behavior
     public $tmpChunkDir = 'file_chunk_tmp';
 
     /**
+     * @var string
+     */
+    public $chunkSizeVar = 'dzchunksize';
+
+    /**
+     * @var string
+     */
+    public $chunkTotalFileSizeVar = 'dztotalfilesize';
+
+    /**
+     * @var string
+     */
+    public $chunkIdVar = 'dzuuid';
+
+    /**
+     * @var string
+     */
+    public $chunkTotalCountVar = 'dztotalchunkcount';
+
+    /**
+     * @var string
+     */
+    public $chunkIndexVar = 'dzchunkindex';
+
+    /**
      * @inheritdoc
      */
     public function init()
     {
         parent::init();
-        if(empty($this->uploadManager)) {
+        if (empty($this->uploadManager)) {
             $this->uploadManager = Yii::$app->upload;
         }
-
-        $this->chunk = boolval(Yii::$app->getRequest()->getBodyParam('dzchunksize', false));
     }
 
     /**
@@ -93,10 +117,10 @@ class FileUploadBehavior extends Behavior
     {
         return [
             ActiveRecord::EVENT_BEFORE_VALIDATE => 'beforeValidate',
-            ActiveRecord::EVENT_BEFORE_INSERT => 'beforeSave',
-            ActiveRecord::EVENT_BEFORE_UPDATE => 'beforeSave',
-            ActiveRecord::EVENT_BEFORE_DELETE => 'beforeDelete',
-            ActiveRecord::EVENT_AFTER_FIND => 'afterFind',
+            ActiveRecord::EVENT_BEFORE_INSERT   => 'beforeSave',
+            ActiveRecord::EVENT_BEFORE_UPDATE   => 'beforeSave',
+            ActiveRecord::EVENT_BEFORE_DELETE   => 'beforeDelete',
+            ActiveRecord::EVENT_AFTER_FIND      => 'afterFind',
         ];
     }
 
@@ -107,9 +131,21 @@ class FileUploadBehavior extends Behavior
 
     public function beforeValidate($event)
     {
+        if (!$this->chunk) {
+            $this->initFileInstance();
+        }
+    }
+
+    /**
+     *
+     */
+    protected function initFileInstance()
+    {
         $model = $this->owner;
         if (!$this->uploadedFile) {
-            $this->uploadedFile = ($this->fileInstanceName === null ? UploadedFile::getInstance($model, $this->attribute) : UploadedFile::getInstanceByName($this->fileInstanceName));
+            $this->uploadedFile = ($this->fileInstanceName === null ?
+                UploadedFile::getInstance($model, $this->attribute) :
+                UploadedFile::getInstanceByName($this->fileInstanceName));
         }
 
         if (!$this->uploadedFile && $model->{$this->attribute} instanceof UploadedFile) {
@@ -118,20 +154,20 @@ class FileUploadBehavior extends Behavior
 
         if ($this->uploadedFile instanceof UploadedFile) {
             $model->{$this->attribute} = $this->uploadedFile;
-            $validator = Validator::createValidator(
+            $validator                 = Validator::createValidator(
                 'file',
                 $model,
                 $this->attribute,
                 [
                     'extensions' => $this->extensions,
-                    'minSize' => $this->minSize,
-                    'maxSize' => $this->maxSize,
+                    'minSize'    => $this->minSize,
+                    'maxSize'    => $this->maxSize,
                 ]
             );
             $validator->validateAttribute($model, $this->attribute);
         }
 
-        if($this->uploadedFile && $this->uploadedFile->name && $this->base_name){
+        if ($this->uploadedFile && $this->uploadedFile->name && $this->base_name) {
             $model->{$this->base_name} = $this->uploadedFile->name;
         }
     }
@@ -142,16 +178,26 @@ class FileUploadBehavior extends Behavior
             return;
         }
 
+        if (!$this->chunk) {
+            $this->deleteByPost();
+
+            if ($this->uploadedFile instanceof UploadedFile) {
+                $this->removeFile();
+                $this->saveFile();
+            }
+        }
+    }
+
+    /**
+     *
+     */
+    protected function deleteByPost()
+    {
         if ($delete = Yii::$app->request->post($this->deleteFileParam)) {
             if (in_array($this->owner->{$this->attribute}, (array)$delete)) {
                 $this->removeFile();
                 $this->owner->{$this->attribute} = null;
             }
-        }
-
-        if ($this->uploadedFile instanceof UploadedFile) {
-            $this->removeFile();
-            $this->saveFile();
         }
     }
 
@@ -160,32 +206,53 @@ class FileUploadBehavior extends Behavior
         $this->uploadManager->removeFile($this->currentFile);
     }
 
+    /**
+     * @throws \yii\web\HttpException
+     */
     protected function saveFile()
     {
         $path = $this->getNewFilePath($this->uploadedFile);
 
-        if ($this->chunk) {
-            $chunkPath = $this->getChunkTmpPath($this->uploadedFile);
-            $result    = $this->uploadManager->saveFileByChunk($this->uploadedFile, $path, $chunkPath);
+        $this->uploadManager->saveFile($this->uploadedFile, $path);
+        $this->owner->{$this->attribute} = $path;
+    }
 
-            if($result === UploadManager::CHUNK_SAVE_DONE) {
-                $this->chunkComplite = true;
-                $this->owner->{$this->attribute} = $path;
+    /**
+     * @return int
+     * @throws \yii\base\Exception
+     * @throws \yii\web\HttpException
+     */
+    protected function saveFileByChunk()
+    {
+        $path      = $this->getNewFilePath($this->uploadedFile);
+        $chunkPath = $this->getChunkTmpPath($this->uploadedFile);
 
-                $this->owner->clearErrors('path');
-            } else {
-                $this->owner->addError('path', 'Chunk process!');
-            }
+        $chunkParams = [];
 
-        } else {
-            $this->uploadManager->saveFile($this->uploadedFile, $path);
-            $this->owner->{$this->attribute} = $path;
+        $chunkParams['totalSize']  = Yii::$app->getRequest()->getBodyParam($this->chunkTotalFileSizeVar);
+        $chunkParams['size']       = Yii::$app->getRequest()->getBodyParam($this->chunkSizeVar);
+        $chunkParams['totalCount'] = Yii::$app->getRequest()->getBodyParam($this->chunkTotalCountVar);
+        $chunkParams['index']      = Yii::$app->getRequest()->getBodyParam($this->chunkIndexVar);
+        array_filter($chunkParams);
+        if (count($chunkParams) < 3) {
+            return UploadManager::CHUNK_SAVE_ERROR;
         }
+
+        $result = $this->uploadManager->saveFileByChunk($this->uploadedFile, $path, $chunkPath, $chunkParams);
+
+        if ($result === UploadManager::CHUNK_SAVE_DONE) {
+            $this->chunkComplite             = true;
+            $this->owner->{$this->attribute} = $path;
+        } else {
+            $this->owner->{$this->attribute} = '';
+        }
+
+        return $result;
     }
 
     protected function getNewFilePath(UploadedFile $uploadedFile)
     {
-        $basePath = $this->getUploadPath();
+        $basePath = '';//$this->getUploadPath();
 
         $uid  = md5(uniqid($basePath));
         $dir1 = substr($uid, 0, 2);
@@ -207,7 +274,7 @@ class FileUploadBehavior extends Behavior
     {
         $runtimePath = Yii::$app->runtimePath;
         $tmpDir      = $this->tmpChunkDir;
-        $fileName    = md5(Yii::$app->getRequest()->getBodyParam('dzuuid'));
+        $fileName    = md5(Yii::$app->getRequest()->getBodyParam($this->chunkIdVar));
         $ext         = $uploadedFile->getExtension();
         $tmpDirPath  = "{$runtimePath}/{$tmpDir}";
 
@@ -255,5 +322,45 @@ class FileUploadBehavior extends Behavior
     public function setUploadedFile(UploadedFile $file)
     {
         $this->uploadedFile = $file;
+    }
+
+    /**
+     * @return boolean
+     * @throws \yii\base\Exception
+     * @throws \yii\web\HttpException
+     */
+    public function saveWithChunk()
+    {
+        $this->chunk = true;
+
+        $this->initFileInstance();
+
+        $this->deleteByPost();
+
+        $result = true;
+        if ($this->uploadedFile instanceof UploadedFile) {
+
+            $chunkResult = $this->saveFileByChunk();
+
+            switch ($chunkResult) {
+                case UploadManager::CHUNK_SAVE_DONE:
+
+                    $result = $this->owner->save();
+                    break;
+                case UploadManager::CHUNK_SAVE_PROCESSING:
+
+                    $result = true;
+                    break;
+                case UploadManager::CHUNK_SAVE_ERROR:
+                    $this->owner->addError($this->attribute, 'Error save file chunk!');
+
+                    $result = false;
+                    break;
+                default:
+                    throw new InvalidArgumentException("Wrong chunk status: {$result}");
+            }
+        }
+
+        return $result;
     }
 }
